Update FilterWarningModal to remove error for users with no query scope.
