+++ src/lib/components/modals/FilterWarning.svelte
@@ -1,17 +1,12 @@
 <script lang="ts">
-  import { removeGenomicFilters, removeUnallowedFilters } from '$lib/stores/Filter.ts';
   import { goto } from '$app/navigation';
-  import {
-    hasGenomicFilter,
-    hasUnallowedFilter,
-    hasInvalidFilter,
-    removeInvalidFilters,
-  } from '$lib/stores/Filter.ts';
   import { getModalStore } from '@skeletonlabs/skeleton';
-  import { panelOpen } from '$lib/stores/SidePanel.ts';
-  let message = '';
-  let backTo = '';
-  let resetQuery = () => {};
+
+  const identity = () => {};
+  $: message = $modalStore[0]?.meta.message || '';
+  $: backTo = $modalStore[0]?.meta.backTo || '';
+  $: resetQuery = $modalStore[0]?.meta.resetQuery || identity;
+
   const modalStore = getModalStore();
   function closedModal() {
     if ($modalStore[0]) {
@@ -22,27 +17,10 @@
     goto(`/${backTo.toLowerCase()}`);
     closedModal();
   }
-  if ($hasGenomicFilter || $hasUnallowedFilter) {
-    message =
-      'Your selected filters contain stigmatizing variables and/or genomic filters, which are not supported with Discover';
-    backTo = 'Explore';
-    resetQuery = () => {
-      panelOpen.set(false);
-      removeGenomicFilters();
-      removeUnallowedFilters();
-      goto(`/discover`);
-      closedModal();
-    };
-  } else if ($hasInvalidFilter) {
-    message =
-      'You are not authorized to access the data in Explore based on your selected filters.';
-    backTo = 'Discover';
-    resetQuery = () => {
-      panelOpen.set(false);
-      removeInvalidFilters();
-      goto(`/explorer`);
-      closedModal();
-    };
+
+  function reset() {
+    resetQuery();
+    closedModal();
   }
 </script>
 
@@ -59,11 +37,10 @@
       <p>Would you like to remove the invalid filters or go back to {backTo}?</p>
       <div>
         <div class="dark">
-          <button
-            class="btn variant-ringed hover:variant-filled-warning"
-            on:click={() => resetQuery()}>Remove Invalid Filters</button
+          <button class="btn variant-ringed hover:variant-filled-warning" on:click={reset}
+            >Remove Invalid Filters</button
           >
-          <button class="btn variant-ringed hover:variant-filled-warning" on:click={() => goBack()}
+          <button class="btn variant-ringed hover:variant-filled-warning" on:click={goBack}
             >Back to {backTo}</button
           >
         </div>
