+++ src/routes/(picsure)/+layout.svelte
@@ -1,4 +1,5 @@
 <script lang="ts">
+  import { onMount } from 'svelte';
   import {
     AppShell,
     Modal,
@@ -6,23 +7,30 @@
     Drawer,
     storePopup,
     type ModalComponent,
+    getModalStore,
+    getDrawerStore,
   } from '@skeletonlabs/skeleton';
   import { computePosition, autoUpdate, offset, shift, flip, arrow } from '@floating-ui/dom';
-  import Navigation from '$lib/components/Navigation.svelte';
-  import { onMount } from 'svelte';
-  import SidePanel from '$lib/components/explorer/results/SidePanel.svelte';
   import { page } from '$app/stores';
+  import { goto, beforeNavigate } from '$app/navigation';
+
+  import { panelOpen } from '$lib/stores/SidePanel.ts';
+  import {
+    hasInvalidFilter,
+    hasGenomicFilter,
+    hasUnallowedFilter,
+    removeGenomicFilters,
+    removeUnallowedFilters,
+    removeInvalidFilters,
+  } from '$lib/stores/Filter.ts';
+
+  import Navigation from '$lib/components/Navigation.svelte';
+  import SidePanel from '$lib/components/explorer/results/SidePanel.svelte';
   import ExportStepper from '$lib/components/explorer/export/ExportStepper.svelte';
   import Footer from '$lib/components/Footer.svelte';
   import ModalWrapper from '$lib/components/modals/ModalWrapper.svelte';
-  import { getModalStore, getDrawerStore } from '@skeletonlabs/skeleton';
-  import { beforeNavigate } from '$app/navigation';
-  import { hasInvalidFilter, hasGenomicFilter, hasUnallowedFilter } from '$lib/stores/Filter.ts';
   import DashboardDrawer from '$lib/components/datatable/DashboardDrawer.svelte';
   import FilterWarning from '$lib/components/modals/FilterWarning.svelte';
-
-  const modalStore = getModalStore();
-  const drawerStore = getDrawerStore();
 
   // Highlight.js
   import hljs from 'highlight.js/lib/core';
@@ -30,6 +38,9 @@
   import python from 'highlight.js/lib/languages/python';
   import { storeHighlightJs } from '@skeletonlabs/skeleton';
   import 'highlight.js/styles/obsidian.css';
+
+  const modalStore = getModalStore();
+  const drawerStore = getDrawerStore();
 
   hljs.registerLanguage('python', python);
   hljs.registerLanguage('r', R);
@@ -59,17 +70,43 @@
     !$page.url.pathname.includes('/distributions');
 
   beforeNavigate(({ to, cancel }) => {
-    if (
-      ($hasInvalidFilter && to?.url.pathname.includes('/explorer')) ||
-      (($hasGenomicFilter || $hasUnallowedFilter) && to?.url.pathname.includes('/discover'))
-    ) {
+    const notAuthorized = to?.url.pathname.includes('/explore') && $hasInvalidFilter;
+    const stigmatizing =
+      to?.url.pathname.includes('/discover') && ($hasGenomicFilter || $hasUnallowedFilter);
+
+    if (stigmatizing || notAuthorized) {
+      let meta = {};
+      if (stigmatizing) {
+        meta = {
+          message:
+            'Your selected filters contain stigmatizing variables and/or genomic filters, which are not supported with Discover',
+          backTo: 'Explore',
+          resetQuery: () => {
+            panelOpen.set(false);
+            removeGenomicFilters();
+            removeUnallowedFilters();
+            goto(`/discover`);
+          },
+        };
+      }
+      if (notAuthorized) {
+        meta = {
+          message:
+            'You are not authorized to access the data in Explore based on your selected filters.',
+          backTo: 'Discover',
+          resetQuery: () => {
+            panelOpen.set(false);
+            removeInvalidFilters();
+            goto(`/explorer`);
+          },
+        };
+      }
+
       cancel();
       modalStore.trigger({
         type: 'component',
         component: 'filterWarning',
-        response: (r: string) => {
-          console.log(r);
-        },
+        meta,
       });
     }
   });
