+++ tests/.auth/noScopeUser.json
@@ -1,0 +1,22 @@
+{
+  "cookies": [],
+  "origins": [
+    {
+      "origin": "http://localhost:4173",
+      "localStorage": [
+        {
+          "name": "consentMode",
+          "value": "{\"ad_storage\":\"denied\",\"analytics_storage\":\"granted\",\"personalization_storage\":\"granted\",\"ad_personalization\":\"denied\",\"ad_data\":\"denied\",\"ad_user_data\":\"denied\"}"
+        },
+        {
+          "name": "user",
+          "value": "{\"uuid\":\"1234\",\"email\":\"test@pic-sure.org\",\"privileges\":[\"PIC_SURE_ANY_QUERY\"],\"acceptedTOS\":true,\"queryTemplate\":{}}"
+        },
+        {
+          "name": "token",
+          "value": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwiZW1haWwiOiJ0ZXN0QHBpYy1zdXJlLm9yZyIsImV4cCI6OTYwOTU3Mjk4MiwiaWF0IjoxNjA5NTcyOTgyfQ.M1W7a3jQNoHQxAUwfj3sDqyVtNH_DkRdzsIF3prIYQA"
+        }
+      ]
+    }
+  ]
+}
